requirements.txt
# Auto-generated requirements.txt
google-genai==1.52.0
python-dotenv==1.0.1
duckduckgo-search==8.1.1
google-search-results==2.4.2
wolframalpha==5.1.3
pandas==2.2.2
numpy==1.24.3
matplotlib==3.9.2
seaborn==0.13.2
plotly==6.0.1
requests==2.32.5
beautifulsoup4==4.12.3
